Create collection item using User fields. Regenerate angular
