+++ packages/devextreme-angular/src/ui/nested/base/index.ts
@@ -50,7 +50,6 @@ export * from './splitter-options';
 export * from './tab-panel-options';
 export * from './text-box-options';
 export * from './text-editor-button-dxi';
-export * from './user-dxi';
 export * from './user';
 export * from './vector-map-projection-config';
 export * from './viz-font';
